+++ src/App.tsx
@@ -44,6 +44,7 @@ import Policies from "./pages/Policies";
 import WebsitePages from "./pages/WebsitePages";
 import BlogPosts from "./pages/BlogPosts";
 import ActivityLog from "./pages/ActivityLog";
+import WebhookLog from "./pages/WebhookLog";
 
 const queryClient = new QueryClient();
 
@@ -92,6 +93,7 @@ const App = () => (
                       <Route path="website-pages" element={<RoleGuard allowed={['admin']}><WebsitePages /></RoleGuard>} />
                       <Route path="blog" element={<RoleGuard allowed={['admin']}><BlogPosts /></RoleGuard>} />
                       <Route path="activity-log" element={<RoleGuard allowed={['admin']}><ActivityLog /></RoleGuard>} />
+                      <Route path="webhook-log" element={<RoleGuard allowed={['admin']}><WebhookLog /></RoleGuard>} />
                       <Route path="profile" element={<Profile />} />
                     </Route>
                     <Route path="/invoice/:id" element={<InvoiceView />} />